Move EvmConfig into submodule
